ValAgent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.1
SQLAlchemy==2.0.36
pydantic==2.12.2
pymysql==1.1.1
python-jose==3.3.0
python-dotenv==1.0.1
email-validator==2.3.0
cryptography==46.0.3
passlib==1.7.4
chromadb==0.5.11
sentence-transformers==2.2.2
huggingface-hub==0.14.1 
tokenizers==0.13.3 
transformers==4.30.2
requests==2.31.0
groq==0.33.0
mcp==1.21.0
streamlit==1.50.0
streamlit-chat==0.1.1
fastapi-mail==1.5.8
httpx==0.28.1

